moved all verilator stuff into legacy/backends/verilator parallel change made to resource files used by verilator
