+++ src/main/scala/chisel3/tester/legacy/backends/verilator/VerilatorBackend.scala
@@ -1,11 +1,11 @@
 // See LICENSE for license details.
 
-package chisel3.tester.experimental.backends.verilator
+package chisel3.tester.legacy.backends.verilator
 
 import chisel3.experimental.{FixedPoint, MultiIOModule}
 import chisel3.internal.firrtl.KnownWidth
-import chisel3.tester.experimental.backends.verilator.Utils.unsignedBigIntToSigned
 import chisel3.tester.internal.{BackendInstance, Context, ThreadedBackend}
+import chisel3.tester.legacy.backends.verilator.Utils.unsignedBigIntToSigned
 import chisel3.tester.{Region, TimeoutException}
 import chisel3.{SInt, _}
